+++ Dfe.Academies.Academisation.Data/Migrations/20230622142127_Remove_OpeningDate_Field.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Dfe.Academies.Academisation.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class Remove_OpeningDate_Field : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "OpeningDate",
+                schema: "academisation",
+                table: "Project");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<DateTime>(
+                name: "OpeningDate",
+                schema: "academisation",
+                table: "Project",
+                type: "datetime2",
+                nullable: true);
+        }
+    }
+}