+++ Dfe.Academies.Academisation.IService/ServiceModels/Legacy/ProjectAggregate/LegacyProjectServiceModel.cs
@@ -21,7 +21,6 @@ namespace Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggre
 		public DateTime? ApplicationReceivedDate { get; init; }
 		public DateTime? AssignedDate { get; init; }
 		public DateTime? HeadTeacherBoardDate { get; init; }
-		public DateTime? OpeningDate { get; init; }
 		public DateTime? BaselineDate { get; init; }
 		public DateTime? LocalAuthorityInformationTemplateSentDate { get; init; }
 		public DateTime? LocalAuthorityInformationTemplateReturnedDate { get; init; }
@@ -66,9 +65,8 @@ namespace Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggre
 		[JsonIgnore] public decimal? PercentageOfGoodOrOutstandingSchoolsInTheDiocesanTrust { get; init; }
 		public decimal? DistanceFromSchoolToTrustHeadquarters { get; init; }
 		public string? DistanceFromSchoolToTrustHeadquartersAdditionalInformation { get; init; }
-		public string? MemberOfParliamentParty { get; init; }
-		public string? MemberOfParliamentName { get; init; }
-		public bool? GeneralInformationSectionComplete { get; init; }
+		public string? MemberOfParliamentNameAndParty { get; init; }
+		public bool? SchoolOverviewSectionComplete { get; init; }
 		public string? SchoolPerformanceAdditionalInformation { get; init; }
 		public string? RationaleForProject { get; init; }
 		public string? RationaleForTrust { get; init; }
@@ -123,7 +121,6 @@ namespace Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggre
 				   Nullable.Equals(ApplicationReceivedDate, other.ApplicationReceivedDate) &&
 				   Nullable.Equals(AssignedDate, other.AssignedDate) &&
 				   Nullable.Equals(HeadTeacherBoardDate, other.HeadTeacherBoardDate) &&
-				   Nullable.Equals(OpeningDate, other.OpeningDate) &&
 				   Nullable.Equals(BaselineDate, other.BaselineDate) &&
 				   Nullable.Equals(LocalAuthorityInformationTemplateSentDate,
 					   other.LocalAuthorityInformationTemplateSentDate) &&
@@ -182,11 +179,9 @@ namespace Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggre
 				   string.Equals(DistanceFromSchoolToTrustHeadquartersAdditionalInformation,
 					   other.DistanceFromSchoolToTrustHeadquartersAdditionalInformation,
 					   StringComparison.InvariantCultureIgnoreCase) &&
-				   string.Equals(MemberOfParliamentParty, other.MemberOfParliamentParty,
+				   string.Equals(MemberOfParliamentNameAndParty, other.MemberOfParliamentNameAndParty,
 					   StringComparison.InvariantCultureIgnoreCase) &&
-				   string.Equals(MemberOfParliamentName, other.MemberOfParliamentName,
-					   StringComparison.InvariantCultureIgnoreCase) &&
-				   GeneralInformationSectionComplete == other.GeneralInformationSectionComplete &&
+				   SchoolOverviewSectionComplete == other.SchoolOverviewSectionComplete &&
 				   string.Equals(SchoolPerformanceAdditionalInformation, other.SchoolPerformanceAdditionalInformation,
 					   StringComparison.InvariantCultureIgnoreCase) &&
 				   string.Equals(RationaleForProject, other.RationaleForProject,
@@ -259,7 +254,6 @@ namespace Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggre
 			hashCode.Add(ApplicationReceivedDate);
 			hashCode.Add(AssignedDate);
 			hashCode.Add(HeadTeacherBoardDate);
-			hashCode.Add(OpeningDate);
 			hashCode.Add(BaselineDate);
 			hashCode.Add(LocalAuthorityInformationTemplateSentDate);
 			hashCode.Add(LocalAuthorityInformationTemplateReturnedDate);
@@ -302,9 +296,8 @@ namespace Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggre
 			hashCode.Add(DistanceFromSchoolToTrustHeadquarters);
 			hashCode.Add(DistanceFromSchoolToTrustHeadquartersAdditionalInformation,
 				StringComparer.InvariantCultureIgnoreCase);
-			hashCode.Add(MemberOfParliamentParty, StringComparer.InvariantCultureIgnoreCase);
-			hashCode.Add(MemberOfParliamentName, StringComparer.InvariantCultureIgnoreCase);
-			hashCode.Add(GeneralInformationSectionComplete);
+			hashCode.Add(MemberOfParliamentNameAndParty, StringComparer.InvariantCultureIgnoreCase);
+			hashCode.Add(SchoolOverviewSectionComplete);
 			hashCode.Add(SchoolPerformanceAdditionalInformation, StringComparer.InvariantCultureIgnoreCase);
 			hashCode.Add(RationaleForProject, StringComparer.InvariantCultureIgnoreCase);
 			hashCode.Add(RationaleForTrust, StringComparer.InvariantCultureIgnoreCase);