+++ Dfe.Academies.Academisation.IService/ServiceModels/Legacy/ProjectAggregate/SponsoredProjectServiceModel.cs
@@ -16,7 +16,7 @@
 	{
 		public string? Name { get; init; }
 		public int Urn { get; init; }
-		public DateTime OpeningDate { get; init; }
+		public DateTime ProposedAcademyOpeningDate { get; init; }
 		public bool PartOfPfiScheme { get; init; }
 		public string? LocalAuthorityName { get; init; }
 		public string? Region{ get; init; }