+++ Dfe.Academies.Academisation.Service/Commands/CypressData/CyAddVoluntaryProjectCommand.cs
@@ -1,0 +1,16 @@
+﻿using Dfe.Academies.Academisation.Core;
+using MediatR;
+
+namespace Dfe.Academies.Academisation.Service.Commands.CypressData
+{
+	/// <summary>
+	///     The cy add voluntary project.
+	/// </summary>
+	public class CyAddVoluntaryProjectCommand : CypressDataCommandAbstractBase, IRequest<CommandResult>
+	{
+		/// <summary>
+		///     Gets a value indicating whether has valid arguments.
+		/// </summary>
+		public override bool HasValidArguments => true;
+	}
+}