+++ terraform/locals.tf
@@ -9,12 +9,15 @@ locals {
   container_command                               = var.container_command
   container_secret_environment_variables          = var.container_secret_environment_variables
   enable_event_hub                                = var.enable_event_hub
+  enable_logstash_consumer                        = var.enable_logstash_consumer
+  eventhub_export_log_analytics_table_names       = var.eventhub_export_log_analytics_table_names
   enable_mssql_database                           = var.enable_mssql_database
   enable_dns_zone                                 = var.enable_dns_zone
   dns_zone_domain_name                            = var.dns_zone_domain_name
   dns_ns_records                                  = var.dns_ns_records
   dns_txt_records                                 = var.dns_txt_records
   enable_cdn_frontdoor                            = var.enable_cdn_frontdoor
+  container_apps_allow_ips_inbound                = var.container_apps_allow_ips_inbound
   cdn_frontdoor_enable_rate_limiting              = var.cdn_frontdoor_enable_rate_limiting
   cdn_frontdoor_rate_limiting_duration_in_minutes = var.cdn_frontdoor_rate_limiting_duration_in_minutes
   cdn_frontdoor_rate_limiting_threshold           = var.cdn_frontdoor_rate_limiting_threshold
@@ -22,6 +25,7 @@ locals {
   cdn_frontdoor_custom_domains                    = var.cdn_frontdoor_custom_domains
   cdn_frontdoor_origin_fqdn_override              = var.cdn_frontdoor_origin_fqdn_override
   cdn_frontdoor_origin_host_header_override       = var.cdn_frontdoor_origin_host_header_override
+  cdn_frontdoor_forwarding_protocol               = var.cdn_frontdoor_forwarding_protocol
   key_vault_access_users                          = toset(var.key_vault_access_users)
   key_vault_access_ipv4                           = var.key_vault_access_ipv4
   tfvars_filename                                 = var.tfvars_filename